Add files via upload
I have worked as a senior school teacher and taught Economics, Commerce and Environmental Science. I worked as an ICSE  examiner for five years. I have worked in residential and non-residential schools. I am also committed to Environment and was in charge of the environmental club of the school. I worked with Rotary club for the benefit of society. I have authored life skill books for primary school.